Project/services/user_management/app_user_management/requirement_game.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==4.2.16
django-allauth==65.2.0
django-cleanup==9.0.0
django-environ==0.11.2
django-htmx==1.21.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==11.0.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
screeninfo==0.8.1
service-identity==24.2.0
setuptools==75.5.0
shortuuid==1.0.13
sqlparse==0.5.1
style==1.1.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
update==0.0.1
whitenoise==6.8.2
zope.interface==7.1.1
